Workaround LED staying off in client mode after enabling the LED again
The ESP SDK LED code seems to be fully event driven. The LED is
constantly on in client mode if a connection is established. If the
LED is turned off and turned on again in this state then it will
stay off. Because the ESP SDK LED code doesn't seem to restore the
LED state to the expected state on re-registering the LED with the
SDK logic.

Workaround this by turning the LED on before registering it to the
SDK logic. This ensures the LED will be on in client mode. This
also works correctly if the SDK logic lets the LED blink. The only
case where this will result in the wrong LED state is when the LED
should actually be off constantly.

Closes #1.